♻ move tests to relevant file
