+++ elements/reigns/src/features/game/Game.test.ts
@@ -8,7 +8,11 @@ import {
 import { getLatestVote } from "../voting/votingSlice";
 import { Game } from "./Game";
 import { mockSdk } from "./mocks";
-import { createCard, createGameState } from "./objectMother";
+import {
+  createCard,
+  createGameDefinition,
+  createGameState,
+} from "./objectMother";
 
 describe("Game", () => {
   beforeEach(() => {
@@ -37,6 +41,79 @@ describe("Game", () => {
       expect(result.phase).toBe(GamePhase.ENDED);
       expect(getSdk().storage.realtime.all(PARTICIPANT_VOTE_TABLE)).toEqual({});
     });
-    it("should clear round resolution", () => {});
+    describe("startGame", () => {
+      it("should select a card", () => {
+        const result = new Game()
+          .startGame(
+            createGameState(
+              createGameDefinition({
+                cards: [createCard({ card: "another card" })],
+              })
+            )
+          )
+          .retrieve();
+        expect(result.selectedCard?.card).toBe("another card");
+      });
+    });
+    describe("answerYes", () => {
+      it("should select a card", () => {
+        const result = new Game()
+          .answerYes(
+            createGameState(
+              createGameDefinition({
+                cards: [createCard({ card: "another card" })],
+              })
+            )
+          )
+          .retrieve();
+        expect(result.selectedCard?.card).toBe("another card");
+      });
+    });
+    describe("answerNo", () => {
+      it("should select a card", () => {
+        const result = new Game()
+          .answerNo(
+            createGameState(
+              createGameDefinition({
+                cards: [createCard({ card: "another card" })],
+              })
+            )
+          )
+          .retrieve();
+        expect(result.selectedCard?.card).toBe("another card");
+      });
+    });
+  });
+  describe("conditions", () => {
+    it("can step through conditions", () => {
+      const card1 = createCard({
+        conditions: "dragon_killed==false",
+        card: "Kill the dragon?",
+        yes_custom: "dragon_killed=true",
+      });
+
+      const card2 = createCard({
+        conditions: "dragon_killed==true",
+        card: "Have a feast?",
+      });
+
+      const gameState = createGameState(
+        createGameDefinition({
+          cards: [card1, card2],
+          deathMessage: "You died",
+          assetsUrl: "test",
+          stats: [{ value: 0, icon: "icon" }],
+        })
+      );
+
+      const turn1 = new Game().startGame(gameState).retrieve();
+      expect(turn1.selectedCard?.card).toBe("Kill the dragon?");
+
+      const turn2 = new Game().answerNo({ ...gameState, ...turn1 }).retrieve();
+      expect(turn2.selectedCard?.card).toBe("Kill the dragon?");
+
+      const turn3 = new Game().answerYes({ ...gameState, ...turn1 }).retrieve();
+      expect(turn3.selectedCard?.card).toBe("Have a feast?");
+    });
   });
 });
